latlngbounds fixed, login checks added

+++ controllers/put_post_flag.py
@@ -25,52 +25,49 @@ from controllers.put_tag import put_tag
 from controllers.get_date_time import get_date_time
 from controllers.get_hash import get_hash
 from controllers.datastore_results import datastore_results
-
-
-
-
 import random
-
-
-
 
 def put_post_flag(self):
     logging.debug("put_post_flag")
-    
-    continue_boolean = False
-    email = ""
-    #try:
-    session = get_current_session()
-    email = session['email']
-    continue_boolean = True
-    #except:
-        #logging.debug("gaesessions exception, do not continue")
-        #continue_boolean = False
-        #show_error_html(self, "session error")
+    if check_login(self):
         
-        
-    if continue_boolean:
-        post_id = self.request.get("post_id")
-
-        filters = {
-            "post_id": post_id,
-        }
-        results, results_exist = datastore_results("Post", filters = filters, inequality_filters = None, order = None, fetch_total = 1, offset = 0, mem_key = None)
-        key = None
-        points = 0
-        for result in results:
-            key = result.key()
-            post_id = result.post_id
-            email = result.email
-            title = result.title
-            tags = result.tags
-            entry = result.entry
-            tags_list = result.tags_list
-            timestamp = result.timestamp
-            points = result.points
-            user_email = session['email']
+        continue_boolean = False
+        email = ""
+        #try:
+        session = get_current_session()
+        email = session['email']
+        continue_boolean = True
+        #except:
+            #logging.debug("gaesessions exception, do not continue")
+            #continue_boolean = False
+            #show_error_html(self, "session error")
             
-            p = model.FlagPost(post_id = post_id, email = email, title = title, tags = tags, entry = entry, tags_list = tags_list, timestamp = timestamp, points=points, added_by = user_email)
-            p.put()
-            db.delete(key)
-        return post_id
+            
+        if continue_boolean:
+            post_id = self.request.get("post_id")
+
+            filters = {
+                "post_id": post_id,
+            }
+            results, results_exist = datastore_results("Post", filters = filters, inequality_filters = None, order = None, fetch_total = 1, offset = 0, mem_key = None)
+            key = None
+            points = 0
+            for result in results:
+                key = result.key()
+                post_id = result.post_id
+                email = result.email
+                title = result.title
+                tags = result.tags
+                entry = result.entry
+                tags_list = result.tags_list
+                timestamp = result.timestamp
+                points = result.points
+                user_email = session['email']
+                
+                p = model.FlagPost(post_id = post_id, email = email, title = title, tags = tags, entry = entry, tags_list = tags_list, timestamp = timestamp, points=points, added_by = user_email)
+                p.put()
+                db.delete(key)
+            return post_id
+            
+    else:
+        self.redirect("/")
